Pre-Service Pack 1
This pack completes the semant method (once the collect_declarations
line is uncommented nothing else should be needed in it), and begins to
fill in the ClassTable at the top. This update also begins to handle
reporting errors by modifying semant_error. See the actual documentation
in the code for more details.